Fixed minor warnings
git-svn-id: https://asteriskvideo.svn.sourceforge.net/svnroot/asteriskvideo@233 57fa1218-b025-0410-bec8-92b503356f72

+++ libh324m/src/H223Muxer.cpp
@@ -239,8 +239,8 @@ inline BYTE H223Muxer::Multiplex()
 					log->SetMuxInfo("endflg");
 				} else {
 					//Create the flag
-					buffer[0] = ~0xE1;
-					buffer[1] = ~0x4D;
+					buffer[0] = (BYTE)~0xE1;
+					buffer[1] = (BYTE)~0x4D;
 					//Log
 					log->SetMuxInfo("dneflg");
 				}
